+++ MHDF-Tools/src/main/java/cn/ChengZhiYa/MHDFTools/command/CommandRegister.java
@@ -26,9 +26,9 @@ import cn.ChengZhiYa.MHDFTools.command.subCommand.misc.time.Day;
 import cn.ChengZhiYa.MHDFTools.command.subCommand.misc.time.Night;
 import cn.ChengZhiYa.MHDFTools.command.subCommand.misc.weather.Sun;
 import cn.ChengZhiYa.MHDFTools.listeners.player.*;
-import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.CraftingTable;
-import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.EnderChest;
-import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.ShulkerBox;
+import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.mode.EnderChest;
+import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.mode.ShulkerBox;
+import cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.util.CraftingTable;
 import cn.ChengZhiYa.MHDFTools.listeners.server.ServerCommandListener;
 import cn.ChengZhiYa.MHDFTools.listeners.server.ServerMOTDListener;
 import cn.ChengZhiYa.MHDFTools.listeners.server.menu.HomeMenu;
