+++ MHDF-Tools/src/main/java/cn/ChengZhiYa/MHDFTools/listeners/player/fastuse/mode/EnderChest.java
@@ -1,4 +1,4 @@
-package cn.ChengZhiYa.MHDFTools.listeners.player.fastuse;
+package cn.ChengZhiYa.MHDFTools.listeners.player.fastuse.mode;
 
 import cn.ChengZhiYa.MHDFTools.MHDFTools;
 import org.bukkit.Material;
@@ -8,15 +8,15 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.player.PlayerInteractEvent;
 
-public final class CraftingTable implements Listener {
+public final class EnderChest implements Listener {
     @EventHandler
     public void onPlayerInteractEvent(PlayerInteractEvent event) {
-        if (MHDFTools.instance.getConfig().getBoolean("FastUseCraftingTableSettings.Enable")) {
+        if (MHDFTools.instance.getConfig().getBoolean("FastUseEnderChestSettings.Enable")) {
             if (event.getAction() == Action.RIGHT_CLICK_AIR) {
                 Player player = event.getPlayer();
-                if (player.hasPermission("MHDFTools.FastUse.CraftingTable")) {
-                    if (player.getInventory().getItemInMainHand().getType() == Material.CRAFTING_TABLE) {
-                        player.openWorkbench(null, true);
+                if (player.hasPermission("MHDFTools.FastUse.EnderChest")) {
+                    if (player.getInventory().getItemInMainHand().getType() == Material.ENDER_CHEST) {
+                        player.openInventory(player.getEnderChest());
                     }
                 }
             }
